remove require dependency loop
